Serttifdikalar kısmı ayarlanması gerceklestirildi

+++ app/Http/Controllers/api/back/sertifikalar/indexController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\api\back\sertifikalar;
 
+use App\Helpers\myHelper;
 use App\Http\Controllers\Controller;
 use App\Models\SertifikaModel;
 use Illuminate\Http\Request;
@@ -9,7 +10,30 @@ use Yajra\DataTables\DataTables;
 
 class indexController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            if (myHelper::yetkiKontrol('sertifikalar',"aktiflik")===false){
+                return redirect()->route('back.home.index')->with(array(
+                    "type" => "error",
+                    "title" => "Hata",
+                    "text" => "Yetkiniz Yok"
+                ));
+            }
+            return $next($request);
+        });
+    }
+
     public function index(){
+
+        if (myHelper::yetkiKontrol('sertifikalar',"listeleme")===false){
+            return redirect()->route('back.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         $query = SertifikaModel::query();
         $data = DataTables::of($query)
             ->addIndexColumn()
@@ -44,6 +68,15 @@ class indexController extends Controller
 
     // EKLEME KISMI AYARLANMASI
     public function store(Request $request){
+
+        if (myHelper::yetkiKontrol('sertifikalar',"ekleme")===false){
+            return redirect()->route('back.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         $data = $request->except("_token");
         $result = SertifikaModel::create($data);
 
@@ -66,16 +99,43 @@ class indexController extends Controller
 
     // GUNCELLEME KISMI AYARLANAMSI
     public function edit(SertifikaModel $item){
+
+        if (myHelper::yetkiKontrol('sertifikalar',"guncelleme")===false){
+            return redirect()->route('back.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         return response()->json($item);
     }
 
     // GORUNTULEME KISMI AYARLANMASI
     public function show(SertifikaModel $item){
+
+        if (myHelper::yetkiKontrol('sertifikalar',"goruntuleme")===false){
+            return redirect()->route('back.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         return response()->json($item);
     }
 
     // GUNCELLEME KISMI
     public function update(Request $request,SertifikaModel $item){
+
+        if (myHelper::yetkiKontrol('sertifikalar',"guncelleme")===false){
+            return redirect()->route('back.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         $data = $request->except("_token");
         $result = $item->update($data);
 
@@ -98,6 +158,15 @@ class indexController extends Controller
 
     // SILME KISMI AYARLANMASI
     public function delete(SertifikaModel $item){
+
+        if (myHelper::yetkiKontrol('sertifikalar',"silme")===false){
+            return redirect()->route('back.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         $sonuc = $item->delete();
 
         if ($sonuc) {
@@ -120,6 +189,14 @@ class indexController extends Controller
     // AKTIF PASIF KISMI AYARLAMASI
     public function isActiveSetter(Request $request, SertifikaModel $item)
     {
+        if (myHelper::yetkiKontrol('sertifikalar',"guncelleme")===false){
+            return redirect()->route('back.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         $data = ($request->data == "true") ? 1 : 0;
         $item->update(array(
             "sr_durum" => $data
@@ -128,6 +205,15 @@ class indexController extends Controller
 
     // SIRALAMA KISMI AYARLANMASI
     public function rankSetter(Request $request){
+
+        if (myHelper::yetkiKontrol('sertifikalar',"guncelleme")===false){
+            return redirect()->route('back.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         parse_str($request->post('data'), $sirala);
         $sirala = $sirala['item'];
